Support Required(Any())
This allows the generated JSON schema to contain `anyOf` constraints, designed to signal
that while two or more fields may be optional independently, it's required to provide at least
one of those optional fields.

One prime example of this are `brightness` and `color` in the `SetLight` command. Neither one
is mandatory, but at least one must be provided otherwise the call does nothing at all.