Support Required(Any()) (#60)
* Support Required(Any())

This allows the generated JSON schema to contain `anyOf` constraints, designed to signal
that while two or more fields may be optional independently, it's required to provide at least
one of those optional fields.

One prime example of this are `brightness` and `color` in the `SetLight` command. Neither one
is mandatory, but at least one must be provided otherwise the call does nothing at all.

* Fix linting problems

* Run `black` to format files properly

* Test that validations work. They didn't. Fixed

* Add test that verifies description is preserved

* Update voluptuous_openapi/__init__.py

Co-authored-by: Allen Porter <allen.porter@gmail.com>

* Ensure that the most specific description prevails over the generic description

* Format with black

---------

Co-authored-by: Allen Porter <allen.porter@gmail.com>